refactor: extract shared instrument-aware Money type to shared/pkg/money
Create shared/pkg/money.Money wrapping shared/platform/quantity.Money
with instrument-aware constructors and backward-compatible accessors.

Update services/current-account/domain/quantity.go to re-export from
shared/pkg/money via type aliases and var constructor re-exports,
maintaining full backward compatibility within the service boundary.

Update tests/integration/balance_ownership_test.go to use sharedmoney
constructors directly instead of routing through cadomain for Money
creation, decoupling the integration test from current-account's
internal money construction helpers.

No import cycles introduced. All service domain tests pass.
