+++ tests/integration/balance_ownership_test.go
@@ -29,6 +29,7 @@ import (
 	cadomain "github.com/meridianhub/meridian/services/current-account/domain"
 	pkdomain "github.com/meridianhub/meridian/services/position-keeping/domain"
 	"github.com/meridianhub/meridian/shared/domain/money"
+	sharedmoney "github.com/meridianhub/meridian/shared/pkg/money"
 	"github.com/meridianhub/meridian/shared/platform/await"
 )
 
@@ -260,8 +261,8 @@ func (m *MockCurrentAccountService) Deposit(ctx context.Context, accountID strin
 	if !ok {
 		return fmt.Errorf("account not found: %s", accountID)
 	}
-	// Convert money.Money to cadomain.Money
-	caAmount, err := cadomain.NewMoneyDecimal(amount.Amount(), cadomain.Currency(amount.Currency().String()))
+	// Convert money.Money to sharedmoney.Money (cadomain.Money = sharedmoney.Money)
+	caAmount, err := sharedmoney.NewFromDecimal(amount.Amount(), sharedmoney.Currency(amount.Currency().String()))
 	if err != nil {
 		return err
 	}
@@ -281,8 +282,8 @@ func (m *MockCurrentAccountService) Withdraw(ctx context.Context, accountID stri
 	if !ok {
 		return fmt.Errorf("account not found: %s", accountID)
 	}
-	// Convert money.Money to cadomain.Money
-	caAmount, err := cadomain.NewMoneyDecimal(amount.Amount(), cadomain.Currency(amount.Currency().String()))
+	// Convert money.Money to sharedmoney.Money (cadomain.Money = sharedmoney.Money)
+	caAmount, err := sharedmoney.NewFromDecimal(amount.Amount(), sharedmoney.Currency(amount.Currency().String()))
 	if err != nil {
 		return err
 	}
@@ -302,7 +303,7 @@ func (m *MockCurrentAccountService) GetBalance(ctx context.Context, accountID st
 	if !ok {
 		return money.Money{}, fmt.Errorf("account not found: %s", accountID)
 	}
-	// Convert cadomain.Money to money.Money
+	// Convert sharedmoney.Money (= cadomain.Money) to the legacy money.Money type
 	caBalance := account.Balance()
 	return money.MustNew(caBalance.Amount(), money.Currency(string(caBalance.Currency()))), nil
 }
